+++ Example/ReusableDemo/TableViewController.swift
@@ -42,20 +42,20 @@ final class TableViewController: UITableViewController {
   override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
     switch indexPath.section {
     case 0:
-      let colorCell = tableView.dequeueReusableCell(indexPath: indexPath) as MySimpleColorCell
+      let colorCell = tableView.dequeueReusableCell(for: indexPath) as MySimpleColorCell
       let red: CGFloat = indexPath.row == 0 ? 1.0 : 0.0
       colorCell.fill(UIColor(red: red, green: 0.0, blue: 1-red, alpha: 1.0))
       return colorCell
     case 1:
-      let textCell = tableView.dequeueReusableCell(indexPath: indexPath) as MyXIBTextCell
+      let textCell = tableView.dequeueReusableCell(for: indexPath) as MyXIBTextCell
       textCell.fill("{section \(indexPath.section), row \(indexPath.row)}")
       return textCell
     case 2:
-      let infoCell = tableView.dequeueReusableCell(indexPath: indexPath) as MyXIBInfoCell
+      let infoCell = tableView.dequeueReusableCell(for: indexPath) as MyXIBInfoCell
       infoCell.fill("InfoCell #\(indexPath.row)", info: "Info #\(indexPath.row)", details: "Details #\(indexPath.row)")
       return infoCell
     case 3:
-      let pathCell = tableView.dequeueReusableCell(indexPath: indexPath) as MyStoryBoardIndexPathCell
+      let pathCell = tableView.dequeueReusableCell(for: indexPath) as MyStoryBoardIndexPathCell
       pathCell.fill(indexPath)
       return pathCell
     default: